continue exploration of BaseData object
